fix my-radars/
